# api/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/tickets/{idOrKey}" : {
      "get" : {
        "description" : "A segment matching ^[A-Z][A-Z0-9]{1,9}-\\d+$ is treated as a project-key/number display key; otherwise it is parsed as a UUID. Returns 404 if no live ticket the caller can see matches.",
        "operationId" : "getTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "idOrKey",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket found"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Missing or invalid access token"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket not found or caller is not a member"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Fetch a single ticket by UUID id or display key (e.g. ENG-42)",
        "tags" : [ "tickets" ]
      }
    },
    "/api/tickets/{id}" : {
      "delete" : {
        "operationId" : "deleteTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Ticket soft-deleted"
          },
          "401" : {
            "description" : "Missing or invalid access token"
          },
          "403" : {
            "description" : "Caller is a member but not an ADMIN"
          },
          "404" : {
            "description" : "Ticket not found or caller is not a member"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Soft-delete a ticket (ADMIN only)",
        "tags" : [ "tickets" ]
      },
      "patch" : {
        "description" : "Null/absent fields are left unchanged; an explicit empty-string description clears it. A present-but-blank title is a 400. Status is NOT changed here (use the transition endpoint). updatedAt always advances; createdAt and number are never changed.",
        "operationId" : "updateTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTicketRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket updated"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Validation error"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Missing or invalid access token"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket not found or caller is not a member"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Partially update a ticket (any member)",
        "tags" : [ "tickets" ]
      }
    },
    "/api/tickets/{id}/labels" : {
      "put" : {
        "description" : "Full idempotent replace: the supplied labelIds become the ticket's complete label set (an empty list clears all). Every id must be a live label in the ticket's own project (else 400). Duplicate ids are de-duplicated.",
        "operationId" : "setTicketLabels",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetTicketLabelsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Labels replaced"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Validation error or cross-project label"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Missing or invalid access token"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket not found or caller is not a member"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Replace a ticket's labels (any member)",
        "tags" : [ "tickets" ]
      }
    },
    "/api/tickets/{id}/transition" : {
      "post" : {
        "description" : "Any status may transition to any other (MVP — no state-machine). A transition to the current status is a 200 no-op and publishes no event.",
        "operationId" : "transitionTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket transitioned (or no-op if already in that status)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Validation error"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Missing or invalid access token"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "Ticket not found or caller is not a member"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Transition a ticket's status (any member)",
        "tags" : [ "tickets" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SetTicketLabelsRequest" : {
        "type" : "object",
        "properties" : {
          "labelIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        },
        "required" : [ "labelIds" ]
      },
      "TicketResponse" : {
        "type" : "object",
        "properties" : {
          "assigneeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "key" : {
            "type" : "string"
          },
          "labelIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "P0", "P1", "P2", "P3" ]
          },
          "projectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reporterId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "IN_REVIEW", "DONE" ]
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "TransitionRequest" : {
        "type" : "object",
        "properties" : {
          "toStatus" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "IN_REVIEW", "DONE" ]
          }
        },
        "required" : [ "toStatus" ]
      },
...
      "UpdateTicketRequest" : {
        "type" : "object",
        "properties" : {
          "assigneeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "description" : {
            "maxLength" : 65536,
            "minLength" : 0,
            "type" : "string"
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "P0", "P1", "P2", "P3" ]
          },
          "title" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}